+++ pyrehau_neasmart/__init__.py
@@ -9,6 +9,7 @@ except ImportError:
 from .const import _PATH_GET, _PATH_POST
 from .heatarea import RehauNeaSmartHeatarea
 from .iodevice import RehauNeaSmartIoDevice
+from .device import RehauNeaSmartDevice
 from .exceptions import RehauNeaSmartError
 
 __author__ = "Jeoffrey Bauvin"
@@ -111,4 +112,12 @@ class RehauNeaSmart(object):
             for child in oneshot.iter(tag='IODEVICE'):
                 devices.append(RehauNeaSmartIoDevice(self, id=child.attrib['nr'], auto_update=self._auto_update))
         return devices
+
+    def device(self):
+        """
+        Returns:
+            RehauNeaSmartDevice object representing the main controller
+        """
+        return RehauNeaSmartDevice(self, auto_update=self._auto_update)
+